meeting_scheduler/requirements.txt
mcp>=1.2.0
google-api-python-client>=2.137.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
python-dateutil>=2.9.0.post0
pytz>=2024.1
google-generativeai>=0.7.2
